hooks: correctly de-duplicate requests for products
My previous implementation here was incorrectly adding the zone ID to
the key-set for the fragments, whivh are zone-agnostic.
